Miscellaneous cleanup.
git-svn-id: http://worldwind31.arc.nasa.gov/svn/trunk@926 822214bb-dcf7-4c8d-a767-17ae0fb20d88

+++ WorldWindIOS/WorldWind/WorldWind/Terrain/WWTerrainTileList.h
@@ -23,7 +23,7 @@
 
 - (WWTerrainTileList*) initWithTessellator:(WWTessellator*)tessellator;
 - (void) addTile:(WWTerrainTile*)tile;
-- (WWTerrainTile*) objectAtIndex:(int)index;
+- (WWTerrainTile*) objectAtIndex:(NSUInteger)index;
 - (NSUInteger) count;
 - (void) beginRendering:(WWDrawContext*)dc;
 - (void) endRendering:(WWDrawContext*)dc;
